chore(acp): split payload renderers
